Event rigging
Properly linked friend classes. Adjusted some members to static.

+++ SDL2Cpp/SDL2Cpp/src/Timer.cpp
@@ -32,7 +32,7 @@ namespace SDL2
 			return 0.0f;
 
 		if(mPaused)
-			return mPausedTicks;
+			return (float)mPausedTicks;
 		else
 			return (float)((SDL_GetTicks() - mStartTicks) / 1000.0f);
 	}
